requirements.txt
fastapi==0.68.*
uvicorn[standard]

tortoise-orm[asyncpg]
